+++ core-game/src/com/giancola/battleship/gamelogic/AI.kt
@@ -1,27 +1,24 @@
 package com.giancola.battleship.gamelogic
 
-import com.badlogic.gdx.math.GridPoint2
 import com.badlogic.gdx.math.MathUtils.random
-import com.giancola.battleship.PlayerData
-import com.giancola.battleship.ShipFactory
-import com.giancola.battleship.ShipPlacement
-import com.giancola.battleship.ShotResult
+import com.giancola.battleship.GameConstants.N_COLS
+import com.giancola.battleship.GameConstants.N_ROWS
 
-class AIPlayer(private val game: GameLogic): GameLogicListener {
+class AIPlayer(private val game: LocalGameLogic): GameLogicListener {
     private val playerData = makeRandomData()
     private val playerId: PlayerId
     private val nRows = playerData.shots.size
     private val nCols = playerData.shots[0].size
 
     init {
-        val regPlayerId = game.registerListener(this, playerData)
+        val regPlayerId = game.registerListener(this)
         require (regPlayerId != null)
         playerId = regPlayerId
     }
 
     companion object {
         fun makeRandomData(): PlayerData {
-            val playerData = PlayerData("Computer", ShipFactory.standardSetup)
+            val playerData = PlayerData("Computer", ShipFactory.standardSetup, N_ROWS, N_COLS)
 
             randomizePlacement(playerData)
 
@@ -31,7 +28,7 @@ class AIPlayer(private val game: GameLogic): GameLogicListener {
         private fun randomizePlacement(playerData: PlayerData) {
             for ((gridX, shipId) in playerData.shipPlacements.keys.withIndex()) {
                 val shipLength = shipId.shipType.width
-                playerData.shipPlacements[shipId] = ShipPlacement(GridPoint2(gridX, 0), GridPoint2(gridX, shipLength-1))
+                playerData.shipPlacements[shipId] = ShipPlacement(Coords(gridX, 0), Coords(gridX, shipLength-1))
             }
         }
     }
@@ -46,23 +43,39 @@ class AIPlayer(private val game: GameLogic): GameLogicListener {
             gridY = random(nCols-1)
         } while (playerData.shots[gridX][gridY])
 
-        game.shoot(playerId, gridX, gridY)
-
+        Thread {
+            randomThink()
+            game.shoot(playerId, gridX, gridY)
+        }.start()
     }
 
 
     //--------------- GameLogicListener methods ---------------
 
-    override fun onGameStarted(whoseTurn: PlayerId) {
+    override fun onGameStarting(playerId: PlayerId) {
+    }
+
+    override fun onGameStarted() {
+        game.setPlacement(playerId, playerData)
+    }
+
+    override fun onCombatStarted(whoseTurn: PlayerId) {
         if (whoseTurn == playerId)
             makeRandomShot()
     }
 
-    override fun onEnemyShot(gridX: Int, gridY: Int, shotResult: ShotResult) {
-        makeRandomShot()
+    override fun onShot(shooter: PlayerId, gridX: Int, gridY: Int, shotResult: ShotResult?) {
+        if (shooter != playerId)
+            makeRandomShot()
     }
 
     override fun onGameFinished(winner: PlayerId) {
         //do nothing
+    }
+
+
+    private fun randomThink() {
+        val sleepInterval = random(1000, 4000).toLong()
+        Thread.sleep(sleepInterval)
     }
 }